requirements.txt
numpy==2.2.3
pandas==2.2.3
protobuf==5.29.3
streamlit==1.42.2
google-generativeai==0.8.4